feat: replace monitor.py with graph_monitor.py
