+++ event-monitor/monitor.py
@@ -1,136 +1,146 @@
 import asyncio
-from web3 import Web3
-import os
+import aiohttp
 import json
 import redis
-from hexbytes import HexBytes
+import os
+import ssl
+from web3 import Web3
+from datetime import datetime
 
-class EventMonitor:
+class GraphEventMonitor:
     def __init__(self):
-        self.w3 = Web3(Web3.HTTPProvider(
-            os.getenv('WEB3_RPC_URL', 'https://base-sepolia.g.alchemy.com/v2/jRDWgvakZFvscXO7eOIZKItOQ_FpnfKd'),
-            request_kwargs={'timeout': 30}
-        ))
-        self.contract_address = Web3.to_checksum_address('0x9fa0da29b88cc1479d28cead5a12ff498528a9d0')
+        self.graph_url = "https://api.studio.thegraph.com/query/12344/onii-chan/version/latest"
         self.redis_client = redis.Redis.from_url(os.getenv('REDIS_URL', 'redis://localhost:6379'))
+        self.w3 = Web3()
+        self.blocks_per_query = 2
+        self.poll_interval = 5
+        
+        # Create SSL context
+        self.ssl_context = ssl.create_default_context()
+        self.ssl_context.check_hostname = False
+        self.ssl_context.verify_mode = ssl.CERT_NONE
 
-        if not self.w3.is_connected():
-            raise Exception("Failed to connect to RPC endpoint")
+    async def make_request(self, query):
+        async with aiohttp.ClientSession() as session:
+            async with session.post(
+                self.graph_url,
+                json={
+                    "query": query,
+                    "operationName": "Subgraphs",
+                    "variables": {}
+                },
+                ssl=self.ssl_context
+            ) as response:
+                return await response.json()
+        
+    async def fetch_tips(self, from_block, to_block):
+        query = """
+        {
+            tips(
+                first: 10, 
+                where: { 
+                    blockNumber_gte: "%d", 
+                    blockNumber_lte: "%d" 
+                }, 
+                orderBy: blockNumber
+            ) {
+                id
+                from
+                tokenId
+                amount
+                tokenAddress
+                message
+                blockNumber
+                blockTimestamp
+                transactionHash
+            }
+        }
+        """ % (from_block, to_block)
 
-        try:
-            current_block = self.w3.eth.block_number
-            print(f"Successfully connected. Current block: {current_block}")
-        except Exception as e:
-            raise Exception(f"Failed to get current block: {str(e)}")
+        result = await self.make_request(query)
+        return result.get('data', {}).get('tips', [])
+
+    async def get_latest_block(self):
+        query = """
+        {
+            _meta {
+                block {
+                    number
+                }
+            }
+        }
+        """
+        
+        result = await self.make_request(query)
+        return int(result.get('data', {}).get('_meta', {}).get('block', {}).get('number', 0))
 
     async def monitor_events(self):
-        print("Starting event monitor...")
+        print("Starting Graph event monitor...")
         stored_block = None
         
         while True:
             try:
-                try:
-                    current_block = self.w3.eth.block_number
-                except Exception as e:
-                    print(f"Error getting block number: {str(e)}")
-                    await asyncio.sleep(5)
-                    continue
-
+                # Get current block from The Graph
+                latest_block = await self.get_latest_block()
+                
                 if stored_block is None:
-                    stored_block = current_block - 1
+                    stored_block = latest_block - 1
                     print(f"Starting from block {stored_block}")
 
-                if stored_block >= current_block:
-                    await asyncio.sleep(5)
+                # Calculate block range for this query
+                from_block = stored_block + 1
+                to_block = min(from_block + self.blocks_per_query - 1, latest_block)
+
+                if from_block > latest_block:
+                    print("Waiting for new blocks...")
+                    await asyncio.sleep(self.poll_interval)
                     continue
 
-                target_block = stored_block + 1
-                print(f"Checking block {target_block}")
-
-                try:
-                    logs = self.w3.eth.get_logs({
-                        'fromBlock': target_block,
-                        'toBlock': target_block,
-                        'address': self.contract_address
-                    })
-                except Exception as e:
-                    print(f"Error getting logs for block {target_block}: {str(e)}")
-                    await asyncio.sleep(5)
-                    continue
-
-                for log in logs:
+                print(f"Checking blocks {from_block} to {to_block}")
+                tips = await self.fetch_tips(from_block, to_block)
+                
+                for tip in tips:
                     try:
-                        await self.process_event(log)
+                        await self.process_event(tip)
                     except Exception as e:
-                        print(f"Error processing event: {str(e)}")
+                        print(f"Error processing tip: {str(e)}")
                         continue
 
-                stored_block = target_block
+                # Update stored block
+                stored_block = to_block
 
             except Exception as e:
                 print(f"Loop error: {str(e)}")
-                await asyncio.sleep(5)
+                await asyncio.sleep(self.poll_interval)
                 continue
 
-            await asyncio.sleep(2)
+            await asyncio.sleep(self.poll_interval)
 
-    def hex_to_address(self, hex_data):
-        """Safely convert hex data to address."""
-        if isinstance(hex_data, HexBytes):
-            hex_str = hex_data.hex()
-        else:
-            hex_str = str(hex_data)
-        
-        # Take the last 40 characters (20 bytes) of the hex string
-        address = '0x' + hex_str[-40:]
-        return Web3.to_checksum_address(address)
-
-    async def process_event(self, event):
+    async def process_event(self, tip):
         try:
-            # Print raw event for debugging
-            print("\nProcessing event:")
-            print(f"Topics: {[t.hex() if isinstance(t, HexBytes) else t for t in event['topics']]}")
-            print(f"Data: {event['data'].hex() if isinstance(event['data'], HexBytes) else event['data']}")
-            
-            # Extract indexed parameters from topics
-            from_address = self.hex_to_address(event['topics'][1])
-            token_id = int(event['topics'][2].hex(), 16)
-            
-            # Decode data field
-            data = event['data'].hex() if isinstance(event['data'], HexBytes) else event['data']
-            if data.startswith('0x'):
-                data = data[2:]
-            
-            # Extract amount (first 32 bytes)
-            amount = int(data[:64], 16)
-            
-            # Extract message from the remaining data
-            try:
-                # Skip token address (32 bytes) and get to string data
-                string_offset = int(data[128:192], 16)  # Get string offset
-                string_start = string_offset * 2  # Convert to hex string position
-                string_length = int(data[string_start:string_start+64], 16)  # Get string length
-                message_start = string_start + 64
-                message = bytes.fromhex(data[message_start:message_start+string_length*2]).decode('utf-8')
-            except Exception as e:
-                print(f"Error decoding message: {e}")
-                message = ""
+            # Convert wei to ETH
+            amount_wei = int(tip['amount'])
+            amount_eth = float(self.w3.from_wei(amount_wei, 'ether'))
+
+            # Convert timestamp to readable format
+            timestamp = datetime.fromtimestamp(int(tip['blockTimestamp']))
 
             # Format event
             tip_event = {
                 "type": "tip",
                 "session_id": "default",
-                "amount": float(self.w3.from_wei(amount, 'ether')),
+                "amount": amount_eth,
                 "currency": "ETH",
-                "token_id": token_id,
-                "message": message
+                "token_id": int(tip['tokenId']),
+                "message": tip['message'],
+                "from_address": get_ens(tip['from'])
             }
 
             print(f"Formatted event: {tip_event}")
 
             # Publish to Redis
             self.redis_client.publish(
-                'vtuber_events',
+                'crypto_tips',
                 json.dumps(tip_event)
             )
             
@@ -138,13 +148,44 @@ class EventMonitor:
 
         except Exception as e:
             print(f"Error in process_event: {str(e)}")
-            print(f"Raw event data: {event}")
+            print(f"Raw tip data: {tip}")
             raise e
+
+import requests
+
+def get_ens(address):
+    # Define the API URL
+    api_url = f"https://ethereum-rest.api.mnemonichq.com/wallets/v1beta2/ens/by_address/{address}"
+    
+    # Set the headers with the API key
+    headers = {
+        'X-API-Key': 'IMAbzFuLRsW9GTkBg4LDL0YV5MwAN7mTsXET5PGSL3oYWiMH'
+    }
+    
+    try:
+        # Make the GET request
+        response = requests.get(api_url, headers=headers)
+        response.raise_for_status()  # Raise an error for bad status codes
+        
+        # Parse the JSON response
+        data = response.json()
+        
+        # Return the 'name' field if it exists
+        if 'entities' in data and data['entities']:
+            for entity in data['entities']:
+                if entity.get('name'):
+                    return entity['name']
+        
+        return address
+    except requests.RequestException as e:
+        print(f"Error during API request: {e}")
+        return None
+
 
 async def main():
     while True:
         try:
-            monitor = EventMonitor()
+            monitor = GraphEventMonitor()
             await monitor.monitor_events()
         except Exception as e:
             print(f"Fatal error, restarting in 5 seconds: {str(e)}")
